Now Login is working and css to
The problem was that i did not have name="username" and the same for the password and that is why  but i used th:field which actually did not work for the UserDetailsService and instead of giving usename="username" it gived empry value. Really good lesson. Now i have to work on the Logout cause it is not working!!!